[#3212] Modify some class name and comment
