+++ api/src/main/java/com/alibaba/nacos/api/exception/NacosException.java
@@ -156,7 +156,7 @@ public class NacosException extends Exception {
     public static final int RESOURCE_NOT_FOUND = -404;
     
     /**
-     * http client error code
+     * http client error code,
      * ome exceptions that occurred when the use the Nacos RestTemplate and Nacos AsyncRestTemplate.
      */
     public static final int HTTP_CLIENT_ERROR_CODE = -500;
