+++ common/src/main/java/com/alibaba/nacos/common/http/client/RestResultResponseHandler.java
@@ -19,24 +19,25 @@ package com.alibaba.nacos.common.http.client;
 import com.alibaba.nacos.common.http.HttpRestResult;
 import com.alibaba.nacos.common.http.param.Header;
 import com.alibaba.nacos.common.model.RestResult;
+import com.alibaba.nacos.common.utils.JacksonUtils;
 
 import java.io.InputStream;
 import java.lang.reflect.Type;
 
 /**
- * RestResult response converter
+ * RestResult response handler,
  * Mainly converter response type as {@link RestResult} type.
  *
  * @author mai.jh
  */
-public class RestResultResponseConverter<T> extends AbstractResponseConverterHandler<T> {
+public class RestResultResponseHandler<T> extends AbstractResponseHandler<T> {
     
     @Override
     @SuppressWarnings("unchecked")
     public HttpRestResult<T> convertResult(HttpClientResponse response, Type responseType) throws Exception {
         final Header headers = response.getHeaders();
         InputStream body = response.getBody();
-        T extractBody = com.alibaba.nacos.common.http.handler.ResponseHandler.convert(body, responseType);
+        T extractBody = JacksonUtils.toObj(body, responseType);
         HttpRestResult<T> httpRestResult = convert((RestResult<T>) extractBody);
         httpRestResult.setHeader(headers);
         return httpRestResult;
